# climbtriage-codex/contracts/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ClimbTriage local keyframe analysis",
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/uploads": {
      "post": {
        "summary": "Create Upload",
        "operationId": "create_upload_v1_uploads_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Upload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/uploads/{ident}": {
      "get": {
        "summary": "Upload Info",
        "operationId": "upload_info_v1_uploads__ident__get",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Append",
        "operationId": "append_v1_uploads__ident__patch",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          },
          {
            "name": "upload-offset",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "title": "Upload-Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete",
        "operationId": "delete_v1_uploads__ident__delete",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs": {
      "post": {
        "summary": "Job",
        "operationId": "job_v1_jobs_post",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Job"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{ident}": {
      "get": {
        "summary": "Get Job",
        "operationId": "get_job_v1_jobs__ident__get",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{ident}/cancel": {
      "post": {
        "summary": "Cancel",
        "operationId": "cancel_v1_jobs__ident__cancel_post",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalysisResult": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "const": 1,
            "title": "Schema Version"
          },
          "holds": {
            "items": {
              "$ref": "#/components/schemas/HoldCandidate"
            },
            "type": "array",
            "title": "Holds"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "warning": {
            "type": "string",
            "title": "Warning"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "schema_version",
          "holds",
          "provenance",
          "warning"
        ],
        "title": "AnalysisResult"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HoldCandidate": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "display_number": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Display Number"
          },
          "kind": {
            "type": "string",
            "enum": [
              "hold",
              "volume"
            ],
            "title": "Kind"
          },
          "parts": {
            "items": {
              "items": {
                "prefixItems": [
                  {
                    "type": "number",
                    "maximum": 1.0,
                    "minimum": 0.0
                  },
                  {
                    "type": "number",
                    "maximum": 1.0,
                    "minimum": 0.0
                  }
                ],
                "type": "array",
                "maxItems": 2,
                "minItems": 2
              },
              "type": "array"
            },
            "type": "array",
            "title": "Parts"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "validity": {
            "type": "string",
            "const": "candidate",
            "title": "Validity"
          },
          "timestamp_us": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Timestamp Us"
          },
          "coordinate_space": {
            "type": "string",
            "const": "upright_video_normalized",
            "title": "Coordinate Space"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "parents": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Parents"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "display_number",
          "kind",
          "parts",
          "confidence",
          "validity",
          "timestamp_us",
          "coordinate_space",
          "provenance",
          "parents"
        ],
        "title": "HoldCandidate"
      },
      "Job": {
        "properties": {
          "upload_id": {
            "type": "string",
            "maxLength": 36,
            "minLength": 36,
            "title": "Upload Id"
          },
          "timestamp_us": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Timestamp Us"
          },
          "coordinate_space": {
            "type": "string",
            "const": "upright_video_normalized",
            "title": "Coordinate Space"
          },
          "consent_to_provider": {
            "type": "boolean",
            "const": true,
            "title": "Consent To Provider"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "upload_id",
          "timestamp_us",
          "coordinate_space",
          "consent_to_provider"
        ],
        "title": "Job"
      },
      "JobInfo": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "phase": {
            "type": "string",
            "title": "Phase"
          },
          "completed_units": {
            "type": "integer",
            "maximum": 3.0,
            "minimum": 0.0,
            "title": "Completed Units"
          },
          "total_units": {
            "type": "integer",
            "const": 3,
            "title": "Total Units"
          },
          "retry_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Retry Count"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "status",
          "phase",
          "completed_units",
          "total_units",
          "retry_count",
          "error",
          "result"
        ],
        "title": "JobInfo"
      },
      "Provenance": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "model_revision": {
            "type": "string",
            "title": "Model Revision"
          },
          "algorithm": {
            "type": "string",
            "title": "Algorithm"
          },
          "config_hash": {
            "type": "string",
            "title": "Config Hash"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "provider",
          "model",
          "model_revision",
          "algorithm",
          "config_hash"
        ],
        "title": "Provenance"
      },
      "Upload": {
        "properties": {
          "sha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Sha256"
          },
          "size_bytes": {
            "type": "integer",
            "maximum": 12582912.0,
            "exclusiveMinimum": 0.0,
            "title": "Size Bytes"
          },
          "content_type": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/png"
            ],
            "title": "Content Type"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sha256",
          "size_bytes",
          "content_type"
        ],
        "title": "Upload"
      },
      "UploadInfo": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset"
          },
          "complete": {
            "type": "boolean",
            "title": "Complete"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "offset",
          "complete"
        ],
        "title": "UploadInfo"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}